+++ Sources/Sentry/Public/SentryTransactionContext.h
@@ -1,3 +1,4 @@
+#import "SentrySampleDecision.h"
 #import "SentrySpanContext.h"
 
 NS_ASSUME_NONNULL_BEGIN
@@ -16,7 +17,7 @@ SENTRY_NO_INIT
 /**
  * Parent sampled
  */
-@property (nonatomic) BOOL parentSampled;
+@property (nonatomic) SentrySampleDecision parentSampled;
 
 /**
  * Init a SentryTransactionContext with given name and set other fields by default
@@ -46,7 +47,7 @@ SENTRY_NO_INIT
                      traceId:(SentryId *)traceId
                       spanId:(SentrySpanId *)spanId
                 parentSpanId:(nullable SentrySpanId *)parentSpanId
-               parentSampled:(BOOL)parentSampled;
+               parentSampled:(SentrySampleDecision)parentSampled;
 
 @end