+++ Sources/Sentry/SentrySpanContext.m
@@ -18,7 +18,7 @@ SentrySpanContext () {
     return [self initWithOperation:operation sampled:false];
 }
 
-- (instancetype)initWithOperation:(NSString *)operation sampled:(BOOL)sampled
+- (instancetype)initWithOperation:(NSString *)operation sampled:(SentrySampleDecision)sampled
 {
     return [self initWithTraceId:[[SentryId alloc] init]
                           spanId:[[SentrySpanId alloc] init]
@@ -31,7 +31,7 @@ SentrySpanContext () {
                          spanId:(SentrySpanId *)spanId
                        parentId:(nullable SentrySpanId *)parentId
                       operation:(NSString *)operation
-                        sampled:(BOOL)sampled
+                        sampled:(SentrySampleDecision)sampled
 {
     if (self = [super init]) {
         _traceId = traceId;
@@ -79,10 +79,14 @@ SentrySpanContext () {
         @"type" : SentrySpanContext.type,
         @"span_id" : self.spanId.sentrySpanIdString,
         @"trace_id" : self.traceId.sentryIdString,
-        @"sampled" : self.sampled ? @"true" : @"false",
         @"tags" : _tags.copy
     }
                                                  .mutableCopy;
+
+    // Since we guard for 'undecided', we'll
+    // either send it if it's 'true' or 'false'.
+    if (self.sampled != kSentrySampleDecisionUndecided)
+        [mutabledictionary setValue:SentrySampleDecisionNames[self.sampled] forKey:@"sampled"];
 
     if (self.operation != nil)
         [mutabledictionary setValue:self.operation forKey:@"op"];